Fix footer problem.
This commit fixes the problem, that the footer is stucked to the left
side of the content. This is solved by including the footer in the
article
element with the content class.

Also this article is converted to a div element since after including
the footer both articles contain basicly the same.